// src/__generated__/_enums.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: H7A0JN5+b0mQwoAKt+bDMTewWSgGIfUmAxL3/0qfzx0j8zDV8zCKiAT8PpCTTkPW0tbotULZuVkCMBUljPNn0A==
 */

/* eslint-disable */
// tslint:disable

type HookType = 'onBranchMerge' | 'onPullRequest' | 'onPullRequestClose' | 'onSlashCommand';
export type {HookType}